+++ CHANGELOG.md
@@ -16,6 +16,11 @@
     - Extended `className` option in a view config to take a `function(data, context)` or `Array.<function(data, context)|string>` beside a string
     - Removed `fallback` view
     - Improved view config error output (view can be overrided by defining custom `config-error` view)
+    - Improved `select` view:
+        - Added selected value state in variants list
+        - Added `resetItem` option to specify a reset item in variants
+        - Fixed `placeholder` is not visible
+        - Fixed size when no item selected and no plahceholder
 
 ## 1.0.0-beta.12 (07-03-2019)